refactor: refacotr layout
