+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Aside/index.js
@@ -1,0 +1,3 @@
+import Aside from './Aside';
+
+export default Aside;
