+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Aside/scss/dark.scss
@@ -1,0 +1,6 @@
+$ice-aside-text-color: #fff !default;
+$ice-aside-sub-text-color: #999 !default;
+
+$theme: 'dark';
+
+@import './base.scss';
