+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Aside/scss/light.scss
@@ -1,0 +1,6 @@
+$ice-aside-text-color: #666 !default;
+$ice-aside-sub-text-color: #666 !default;
+
+$theme: 'light';
+
+@import './base.scss';
