+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Footer/Footer.jsx
@@ -1,19 +1,13 @@
 import React, { PureComponent } from 'react';
 import Layout from '@icedesign/layout';
-import cx from 'classnames';
 import Logo from '../Logo';
+
+import './Footer.scss';
 
 export default class Footer extends PureComponent {
   render() {
-    const { className, style } = this.props;
     return (
-      <Layout.Footer
-        className={cx('ice-design-layout-footer', className)}
-        style={{
-          ...style,
-          lineHeight: '36px',
-        }}
-      >
+      <Layout.Footer className="ice-design-layout-footer">
         <div className="ice-design-layout-footer-body">
           <div style={{ filter: 'grayscale(100%)', opacity: 0.3 }}>
             <Logo />
