+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Header/scss/dark.scss
@@ -1,0 +1,6 @@
+$ice-header-text-color: #fff !default;
+$ice-header-icon-down-color: #fff !default;
+
+$theme: 'dark';
+
+@import './base.scss';
