+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Header/scss/light.scss
@@ -1,0 +1,6 @@
+$ice-header-text-color: #666 !default;
+$ice-header-icon-down-color: #666 !default;
+
+$theme: 'light';
+
+@import './base.scss';
