+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Logo/Logo.jsx
@@ -1,6 +1,8 @@
 import React, { PureComponent } from 'react';
 import { Link } from 'react-router-dom';
 
+import './Logo.scss';
+
 export default class Logo extends PureComponent {
   render() {
     return (
