+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/components/Logo/index.js
