+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/scss/dark.scss
@@ -1,36 +1,6 @@
-@import '~@icedesign/skin/variables.scss';
-
-// Basic color variables
-$custom-layout-bg: #2e323f;
-
-// Layout dark styles
-$ice-body-bg: $color-white !default;
-$ice-layout-bg: $custom-layout-bg !default;
-$ice-toggle-btn-bg: $color-brand1-5 !default;
-$ice-toggle-btn-text-color: $color-white !default;
-$ice-logo-text-color: $color-brand1-5 !default;
-$ice-icon-down-color: $color-white !default;
-
-$ice-header-menu-bg: $color-transparent !default;
-$ice-header-text-color: $color-white !default;
-$ice-header-text-hover-color: $color-brand1-5 !default;
-
-$ice-aside-bg: $color-transparent !default;
-$ice-aside-menu-bg: $color-transparent !default;
-$ice-aside-text-color: $color-white !default;
-$ice-aside-sub-text-color: $color-text1-2 !default;
-$ice-aside-text-hover-color: $color-white !default;
-$ice-aside-menu-item-selected-gradient-left-bg: $color-brand1-9 !default;
-$ice-aside-menu-item-selected-gradient-right-bg: $color-brand1-5 !default;
-$ice-aside-menu-item-selected-text-color: $color-white !default;
-$ice-aside-menu-submenu-item-hover-bg: $color-transparent !default;
-$ice-aside-menu-submenu-vertical-bg: $color-white !default;
-
-$ice-copyright-text-color: $color-text1-3 !default;
-$ice-copyright-text-hover-color: $color-brand1-5 !default;
-
-.ice-design-header-aside-footer-layout-dark.ice-design-layout {
-  background-image: url('./images/dark_bg.png');
+.ice-design-layout-dark.ice-design-layout {
+  background-image: url(https://img.alicdn.com/tfs/TB1_YvwXTtYBeNjy1XdXXXXyVXa-1680-370.png);
+  background-color: #2e323f;
 }
 
 $theme: 'dark';
