+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/scss/index.scss
@@ -1,0 +1,2 @@
+@import './dark.scss';
+@import './light.scss';
