+++ react-materials/scaffolds/ice-design-pro/src/layouts/BasicLayout/scss/light.scss
@@ -1,40 +1,6 @@
-@import '~@icedesign/skin/variables.scss';
-
-// Basic color variables
-$custom-layout-bg: #fafafa;
-
-// Layout light styles
-$ice-body-bg: $color-white !default;
-$ice-layout-bg: $custom-layout-bg !default;
-$ice-toggle-btn-bg: $color-brand1-5 !default;
-$ice-toggle-btn-text-color: $color-white !default;
-$ice-collapse-btn-color: $color-text1-3 !default;
-$ice-logo-text-color: $color-brand1-5 !default;
-$ice-icon-down-color: $color-text1-3 !default;
-
-$ice-header-menu-bg: transparent !default;
-$ice-header-text-color: $color-text1-3 !default;
-$ice-header-text-hover-color: $color-brand1-5 !default;
-
-$ice-aside-bg: transparent !default;
-$ice-aside-menu-bg: transparent !default;
-$ice-aside-text-color: $color-text1-3 !default;
-$ice-aside-sub-text-color: $color-text1-3 !default;
-$ice-aside-text-hover-color: $color-white !default;
-$ice-aside-menu-item-selected-gradient-left-bg: $color-brand1-9 !default;
-$ice-aside-menu-item-selected-gradient-right-bg: $color-brand1-5 !default;
-$ice-aside-menu-item-selected-text-color: $color-white !default;
-$ice-aside-menu-submenu-item-hover-bg: transparent !default;
-$ice-aside-menu-submenu-vertical-bg: $color-white !default;
-$ice-aside-menu-collapse-bg: transparent !default;
-$ice-aside-menu-collapse-selected-text-color: $color-brand1-5 !default;
-$ice-aside-collapse-menu-text-color: $color-text1-3 !default;
-
-$ice-copyright-text-color: $color-text1-3 !default;
-$ice-copyright-text-hover-color: $color-brand1-5 !default;
-
-.ice-design-header-aside-footer-layout-light.ice-design-layout {
-  background-image: url('./images/light_bg.png');
+.ice-design-layout-light.ice-design-layout {
+  background-image: url(https://img.alicdn.com/tfs/TB1GsHCd8fH8KJjy1XbXXbLdXXa-1680-370.png);
+  background-color: #fafafa;
 }
 
 $theme: 'light';
