--- react-materials/scaffolds/ice-design-pro/src/layouts/UserLayout/components/Footer/index.js
@@ -1,3 +1,0 @@
-import Footer from './Footer';
-
-export default Footer;
