+++ react-materials/scaffolds/ice-design-pro/src/store/userLogout/action.js
@@ -47,10 +47,9 @@ export const userLogout = () => {
       dispatch(userLogoutSuccess(response.data));
 
       if (response.data.status === 200) {
+        Feedback.toast.success('已登出');
         setAuthority(response.data.currentAuthority);
-
         reloadAuthorized();
-
         dispatch(push('/user/login'));
       } else {
         Feedback.toast.error('请稍后重试');
